gcd+egcd: fix exponential complexity of binary GCD
Oops! The comment saying the bitlen was decreasing at each iteration was
wrong. If the lowest of the two numbers was even when exiting the first
loop, then we'd keep doing subtractions _not_ followed by a shift until
the lowest number stops being even, which may not happen until the very
end. (For some reason I had convinced myself that if the two numbers
were not odd initially, they would be after the first iteration, but
that was a brain bug of mine.)

An example of terrible perfomance was GCD(n, 2) when n is any large odd
number. In that case the algorithm would just to O(n) subtractions,
which is exponential in bitlen(n).

We really want both a and b to be odd when entering the final loop.